Create GpiTestDialog.tsx
